server/game/mixins.py
from rest_framework import status
from rest_framework.response import Response


class DestroyMixin:
    def delete(self, request, *args, **kwargs):
        instance = self.get_object()
        deleted_id = instance.id
        self.perform_destroy(instance)
        return Response(deleted_id, status=status.HTTP_204_NO_CONTENT)

    def perform_destroy(self, instance):
        instance.delete()
