server/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-countries==7.6.1
django-phone-field==1.8.1
django-vite==3.0.4
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.0
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-webpack-boilerplate==1.0.3
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
